app/requirements.txt
Django==3.1.5
gunicorn==20.0.4
psycopg2
django-storages
boto3
geopy
django-behave